+++ src/main/kotlin/dev/slint/ideaplugin/ide/lsp/SlintLspServerDescriptor.kt
@@ -1,7 +1,11 @@
 package dev.slint.ideaplugin.ide.lsp
 
 import com.intellij.execution.configurations.GeneralCommandLine
+import com.intellij.execution.process.OSProcessHandler
+import com.intellij.execution.process.ProcessAdapter
+import com.intellij.execution.process.ProcessEvent
 import com.intellij.ide.plugins.PluginManager
+import com.intellij.openapi.components.service
 import com.intellij.openapi.extensions.PluginId
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
@@ -17,98 +21,23 @@ import dev.slint.ideaplugin.lang.SlintFileType
 import org.eclipse.lsp4j.services.LanguageServer
 import com.intellij.openapi.util.SystemInfo
 import dev.slint.ideaplugin.SlintBundle
+import dev.slint.ideaplugin.ide.services.SlintServerService
 import java.nio.file.Path
+import java.util.*
+import kotlin.concurrent.schedule
 
 @Suppress("UnstableApiUsage")
 class SlintLspServerDescriptor(project: Project) : ProjectWideLspServerDescriptor(project, "Slint") {
 
     override fun isSupportedFile(file: VirtualFile) = file.fileType == SlintFileType
 
-    override fun createCommandLine(): GeneralCommandLine {
-        val settingState = SlintSettingsState.getInstance().lspSettings
-
-        val parameters = mutableListOf<String>()
-        if (settingState.args.isNotEmpty()) {
-            val args = settingState.args.split("\\s+".toRegex())
-            parameters.addAll(args)
-        }
-
-        if (settingState.includePaths.isNotEmpty()) {
-            parameters.add("-I")
-            settingState.includePaths.forEach {
-                parameters.add("'${it}'")
-            }
-        }
-
-        if (settingState.backend != SlintBackend.DEFAULT) {
-            parameters.add("--backend")
-            parameters.add(settingState.backend.toString())
-        }
-
-        if (settingState.style != SlintStyle.DEFAULT) {
-            parameters.add("--style")
-            parameters.add(settingState.style.toString())
-        }
-
-        if (settingState.noToolbar) {
-            parameters.add("--no-toolbar")
-        }
-
-        val path = if (settingState.useExternalLsp) {
-            settingState.path
-        } else {
-            getEmbeddedLspPath().toString()
-        }
-
-        return GeneralCommandLine(path).apply {
-            addParameters(parameters)
-            withParentEnvironmentType(GeneralCommandLine.ParentEnvironmentType.CONSOLE)
-            withCharset(Charsets.UTF_8)
-        }
-    }
-
-    private fun getEmbeddedLspPath(): Path? {
-        val pluginPath = PluginManager
-            .getInstance()
-            .findEnabledPlugin(PluginId.getId(dev.slint.ideaplugin.SLINT_PLUGIN_ID))
-            ?.pluginPath
-
-        val programName: String
-
-        if (SystemInfo.isMac) {
-            programName = "Slint Live Preview.app/Contents/MacOS/slint-lsp"
-        } else if (SystemInfo.isLinux) {
-            programName = when (SystemInfo.OS_ARCH) {
-                "x64" -> {
-                    "slint-lsp-x86_64-unknown-linux-gnu"
-                }
-
-                "arm" -> {
-                    "slint-lsp-armv7-unknown-linux-gnueabihf"
-                }
-
-                "arm64" -> {
-                    "slint-lsp-aarch64-unknown-linux-gnu"
-                }
-
-                else -> {
-                    return null
-                }
-            }
-        } else if (SystemInfo.isWindows) {
-            programName = "slint-lsp-x86_64-pc-windows-msvc.exe"
-        } else {
-            return null
-        }
-
-        return pluginPath
-            ?.resolve("language-server/bin")
-            ?.resolve(programName)
-    }
+    override fun createCommandLine(): GeneralCommandLine = CommandLineHandler.createCommandLine()
 
     override fun createInitializationOptions(): Any = SlintSettingsState.getInstance().lspSettings
 
     override fun createLsp4jClient(handler: LspServerNotificationsHandler): Lsp4jClient = LspLanguageClient(handler, project)
+
+    override fun startServerProcess(): OSProcessHandler = ServerProcessHandler.addListeners(super.startServerProcess(), project)
 
     override val lsp4jServerClass: Class<out LanguageServer> = SlintLanguageServer::class.java
     override val lspServerListener: LspServerListener = SlintLspServerListener(project)